server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.12.0",
    "chardet>=5.2.0",
    "fastapi>=0.128.7",
    "jsonschema>=4.26.0",
    "lxml>=5.0.0",
    "openpyxl>=3.1.5",
    "playwright>=1.49.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "jinja2>=3.1",
    "plotly>=5.18",
    "weasyprint>=68.0",
    "python-docx>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "polyfactory>=2.0",
    "aiosqlite>=0.20",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
